# sdks/sdkwork-catalog-app-sdk/openapi/sdkwork-catalog-app-api.sdkgen.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Catalog App API",
    "version": "1.0.0",
    "description": "Typed app API for active catalog browsing, cart management, and user delivery addresses.",
    "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
    "x-sdkwork-sdk-family": "sdkwork-catalog-app-sdk",
    "x-sdkwork-owner": "sdkwork-catalog"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
    "/app/v3/api/catalog/attributes": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List active catalog attributes.",
        "operationId": "attributes.list",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Catalog attribute page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogAttributeListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.attributes",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-pagination-mode": "offset",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
    "/app/v3/api/catalog/categories": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List active catalog categories.",
        "operationId": "categories.list",
        "parameters": [
          {
            "name": "parent_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Catalog category page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogCategoryListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.categories",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-pagination-mode": "offset",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
    "/app/v3/api/catalog/categories/{categoryId}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Retrieve an active catalog category.",
        "operationId": "categories.retrieve",
        "parameters": [
          {
            "$ref": "#/components/parameters/CategoryId"
          }
        ],
        "responses": {
          "200": {
            "description": "Catalog category.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogCategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.categories",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
    "/app/v3/api/catalog/products": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List active catalog products.",
        "operationId": "products.list",
        "parameters": [
          {
            "name": "shop_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "product_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Catalog product page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogProductListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.products",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-pagination-mode": "offset",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
    "/app/v3/api/catalog/products/{productId}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Retrieve an active catalog product.",
        "operationId": "products.retrieve",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          }
        ],
        "responses": {
          "200": {
            "description": "Catalog product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.products",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
    "/app/v3/api/catalog/products/{productId}/skus": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List active SKUs for a catalog product.",
        "operationId": "products.skus.list",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Catalog SKU page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogSkuListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-owner": "sdkwork-catalog",
        "x-sdkwork-api-authority": "sdkwork-catalog-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "catalog.products.skus",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-pagination-mode": "offset",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "PageSize": {
        "name": "page_size",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 20
        }
      },
      "CategoryId": {
        "name": "categoryId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40001,
            "maximum": 79999
          },
          "traceId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CatalogAttributeListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/CatalogAttributePageData"
              }
            }
          }
        ]
      },
      "CatalogCategoryListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/CatalogCategoryPageData"
              }
            }
          }
        ]
      },
...
    }
  },
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ]
}